@@ -9,7 +9,7 @@
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "lint": "echo \"No linting configured\""
+    "lint": "eslint --ext .js,.vue src"
   },
   "keywords": [],
   "author": {
@@ -18,16 +18,35 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@electron-forge/cli": "^7.2.0",
-    "@electron-forge/maker-deb": "^7.2.0",
-    "@electron-forge/maker-rpm": "^7.2.0",
-    "@electron-forge/maker-squirrel": "^7.2.0",
-    "@electron-forge/maker-zip": "^7.2.0",
-    "@electron-forge/plugin-auto-unpack-natives": "^7.2.0",
-    "@electron-forge/plugin-vite": "^7.2.0",
-    "electron": "28.2.2"
+    "@electron-forge/cli": "^7.3.0",
+    "@electron-forge/maker-deb": "^7.3.0",
+    "@electron-forge/maker-rpm": "^7.3.0",
+    "@electron-forge/maker-squirrel": "^7.3.0",
+    "@electron-forge/maker-zip": "^7.3.0",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.3.0",
+    "@electron-forge/plugin-fuses": "^7.3.0",
+    "@electron-forge/plugin-vite": "^7.3.0",
+    "@electron/fuses": "^1.7.0",
+    "@faker-js/faker": "^8.4.1",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "electron": "29.1.0",
+    "eslint": "^8.57.0",
+    "eslint-config-standard-with-typescript": "^43.0.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.22.0",
+    "typescript": "^5.4.2",
+    "vite": "^5.1.5"
   },
   "dependencies": {
-    "electron-squirrel-startup": "^1.0.0"
+    "@mdi/font": "^7.4.47",
+    "electron-squirrel-startup": "^1.0.0",
+    "mysql": "^2.18.1",
+    "mysql2": "^3.9.2",
+    "v-mask": "^2.3.0",
+    "vue": "^3.4.21",
+    "vuetify": "^3.5.8"
   }
 }
