@@ -1,5 +1,6 @@
-const { app, BrowserWindow } = require('electron');
-const path = require('path');
+import { app, BrowserWindow, ipcMain }  from "electron";
+import path from "path"
+import queries  from "../queries";
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require('electron-squirrel-startup')) {
@@ -16,6 +17,68 @@ const createWindow = () => {
     },
   });
 
+  // ADD NEW BORROWER
+  ipcMain.handle('db:addBorrower', async (evt, borrowerData) => {
+    try {
+    
+     const borrowers = await queries.addBorrower(borrowerData.firstName, borrowerData.lastName, borrowerData.email, borrowerData.phone, borrowerData.address);
+      console.log(`Borrower added successfully: ${borrowerData.firstName} ${borrowerData.lastName}`);
+      return borrowers 
+    } catch (err) {
+      console.error('Error adding borrower:', err);
+      return Promise.reject(err); // Send error to frontend
+    }
+  });
+  
+  // ADD NEW LOAN AND UPDATE ACCOUNT BALANCE
+  ipcMain.handle('db:addLoan', async (evt, loanData) => {
+    try {
+      const loan = await queries.addLoan(loanData.borrowerID, loanData.loanAmount, loanData.loanDate);
+      console.log(`Loan added successfully for borrower ${loanData.borrowerID}`);
+      return loan
+    } catch (err) {
+      console.error('Error adding loan:', err);
+      return Promise.reject(err); // Send error to frontend
+    }
+  });
+  
+  // ADD NEW PAYMENT AND UPDATE ACCOUNT BALANCE
+  ipcMain.handle('db:addPayment', async (evt, paymentData) => {
+    try {
+      const payment = await queries.addPayment(paymentData.borrowerID, paymentData.paymentAmount, paymentData.paymentDate);
+      console.log(`Payment added successfully for borrower ${paymentData.borrowerID}`);
+      return payment;
+    } catch (err) {
+      console.error('Error adding payment:', err);
+      return Promise.reject(err); // Send error to frontend
+    }
+  });
+  
+  // GET ALL BORROWERS AND THEIR ACCOUNT BALANCE
+  ipcMain.handle('db:getAllBorrowers', async () => {
+    try {
+      const borrowers = await queries.getAllBorrowers();
+      return borrowers;
+    } catch (err) {
+      console.error('Error getting borrowers:', err);
+      return Promise.reject(err);
+    }
+  });
+  
+  // GET ALL PAYMENTS MADE BY A SPECIFIED BORROWER
+  ipcMain.handle('db:getAllPaymentsByBorrower', async (evt, borrowerID) => {
+    try {
+      const payments = await queries.getAllPaymentsByBorrower(borrowerID);
+      return payments;
+    } catch (err) {
+      console.error('Error getting payments for borrower:', err);
+      return Promise.reject(err); 
+    }
+  });
+
+
+
+  
   // and load the index.html of the app.
   if (MAIN_WINDOW_VITE_DEV_SERVER_URL) {
     mainWindow.loadURL(MAIN_WINDOW_VITE_DEV_SERVER_URL);
